command: switched /leave to thread-based
